add gateway persistence
